feat(app); implement tile component in company selection screen
